+++ providers/capa/standard/capa_test.go
@@ -5,10 +5,10 @@ import (
 
 	. "github.com/onsi/ginkgo/v2" //nolint:staticcheck
 
-	"github.com/giantswarm/cluster-test-suites/v4/internal/common"
-	"github.com/giantswarm/cluster-test-suites/v4/internal/ecr"
-	"github.com/giantswarm/cluster-test-suites/v4/internal/state"
-	"github.com/giantswarm/cluster-test-suites/v4/internal/timeout"
+	"github.com/giantswarm/cluster-test-suites/v5/internal/common"
+	"github.com/giantswarm/cluster-test-suites/v5/internal/ecr"
+	"github.com/giantswarm/cluster-test-suites/v5/internal/state"
+	"github.com/giantswarm/cluster-test-suites/v5/internal/timeout"
 )
 
 var _ = Describe("Common tests", func() {