+++ providers/capz/standard/capz_suite_test.go
@@ -8,7 +8,7 @@ import (
 
 	"github.com/giantswarm/cluster-standup-teardown/v4/pkg/clusterbuilder/providers/capz"
 
-	"github.com/giantswarm/cluster-test-suites/v4/internal/suite"
+	"github.com/giantswarm/cluster-test-suites/v5/internal/suite"
 )
 
 func TestCAPZStandard(t *testing.T) {